+++ examples/music_demo/src/lib.rs
@@ -24,8 +24,8 @@ impl Game for MusicDemo {
 
     fn draw(&self, gfx: &mut Graphics) {
         gfx.clear(Color::BLACK);
-        gfx.print("music demo", 44, 8, Color::WHITE);
-        gfx.print("z: play  x: stop", 32, 116, Color::LIGHT_GREY);
+        gfx.print("music demo", 44.0, 8.0, Color::WHITE);
+        gfx.print("z: play  x: stop", 32.0, 116.0, Color::LIGHT_GREY);
         // Dancing bars while playing, flat while stopped.
         for i in 0..16 {
             let phase = (self.t as f32 / 4.0 + i as f32) % 8.0;
@@ -37,10 +37,10 @@ impl Game for MusicDemo {
                 4
             };
             let c = Color::from_index(8 + (i % 8) as u8);
-            gfx.rect_fill(8 + i * 7, 96 - h, 5, h, c);
+            gfx.rect_fill((8 + i * 7) as f32, (96 - h) as f32, 5.0, h as f32, c);
         }
         if self.playing {
-            gfx.print("now playing: pattern 0", 20, 40, Color::GREEN);
+            gfx.print("now playing: pattern 0", 20.0, 40.0, Color::GREEN);
         }
     }
 }
